fix(review): integrate headless lifecycle split
