+++ src/autoskillit/execution/headless/_headless_recovery.py
@@ -3,6 +3,8 @@
 from __future__ import annotations
 
 import dataclasses
+import json
+import os
 from collections.abc import Sequence
 from pathlib import Path
 from typing import TYPE_CHECKING, NamedTuple
@@ -11,15 +13,20 @@ import regex as re
 
 from autoskillit.core import (
     CliSubtype,
+    extract_bash_write_targets,
     get_logger,
 )
-from autoskillit.execution.headless._headless_path_tokens import _RECOVERABLE_PATH_TOKENS
+from autoskillit.execution.headless._headless_path_tokens import (
+    _RECOVERABLE_PATH_TOKENS,
+    _is_path_outside_cwd,
+)
 from autoskillit.execution.process import _marker_is_standalone
 from autoskillit.execution.session import (
     ClaudeSessionResult,
     _check_expected_patterns,
 )
 from autoskillit.execution.session._session_content import _normalize_model_output
+from autoskillit.execution.session._session_model import _is_parent_assistant_record
 
 if TYPE_CHECKING:
     from autoskillit.core import ResultParser
@@ -77,6 +84,60 @@ def _is_path_capture_pattern(pattern: str) -> str | None:
     if "=" not in pattern[m.end() :]:
         return None
     return token_name
+
+
+def _scan_jsonl_write_paths(
+    stdout: str,
+    cwd: str,
+    *,
+    write_tool_names: frozenset[str] = frozenset({"Write", "Edit"}),
+    bash_tool_name: str = "Bash",
+) -> list[str]:
+    """Scan raw JSONL stdout for Write/Edit/Bash tool calls outside cwd."""
+    if not stdout.strip() or not cwd or not os.path.isabs(cwd):
+        return []
+
+    warnings: list[str] = []
+    for raw_line in stdout.strip().splitlines():
+        line = raw_line.strip()
+        if not line:
+            continue
+        try:
+            obj = json.loads(line)
+        except json.JSONDecodeError:
+            continue
+        if not isinstance(obj, dict) or not _is_parent_assistant_record(obj):
+            continue
+        msg = obj.get("message")
+        if not isinstance(msg, dict):
+            continue
+        content = msg.get("content", [])
+        if not isinstance(content, list):
+            continue
+        for block in content:
+            if not isinstance(block, dict) or block.get("type") != "tool_use":
+                continue
+            tool_name = block.get("name", "")
+            inputs = block.get("input") or {}
+            if not isinstance(inputs, dict):
+                continue
+            if tool_name == bash_tool_name:
+                command = inputs.get("command", "")
+                if isinstance(command, str):
+                    for path in extract_bash_write_targets(command, cwd):
+                        if _is_path_outside_cwd(path, cwd):
+                            normalized = os.path.normpath(path)
+                            warnings.append(
+                                f"Bash command contained write target '{normalized}'"
+                                f" outside session cwd '{cwd}'"
+                            )
+            elif tool_name in write_tool_names:
+                file_path = inputs.get("file_path", "")
+                if isinstance(file_path, str) and _is_path_outside_cwd(file_path, cwd):
+                    warnings.append(
+                        f"{tool_name} tool targeted '{file_path}' outside session cwd '{cwd}'"
+                    )
+    return warnings
 
 
 def _recover_from_separate_marker(
